[ffmpeg] Fix compilation on MacOS
Meson is quite stupid in understanding what language the library is
being built. And the linker flags are selected in a per-language
basis. On MacOS libavfilter has a few objc files, which makes
meson be silently confused and rejecting all the link flags for
the library (even though it continues to use `cc` for the linker).

Personally, I would prefer Meson to error out on such ambiguities,
it would save me two(!) days of debugging time.