[ext_googleangle] D3D11: Fix uniform setting for mat3 followed by float
A backported patch from upstream that fixes window rendering on Qt6:

A single float following a float3x3 will pack tightly in the last row
of the matrix. The uniform setting code for matricies would write a
full row stride's worth of data for the last matrix row, overwriting
the float packed at the end.

Update ExpandMatrix to only write colsSrc worth of values in the last
row.