[ffmpeg] Fix detection of iconv library on MacOS
'iconv' library is a bit weird, since on some platforms it is a part
of 'libc' (on Linux) and on others it is available using a separate
library. FFmpeg has 'iconv' detected in two places, first in 'libraries'
section and second in 'header-func' section. They were conflicting
with each other, breaking the build.