Always set an objective-c compiler with crosscompile information for meson
meson setup is run inside an arch wrapper, however on some envs this is
ignored causing the default meson command line to be used be ran from
native arch. For objective-c code this means all code is built as native
arch causing errors when trying to link to different arch static binaries.